优化代码

+++ src/main/kotlin/app/termora/account/PullService.kt
@@ -4,6 +4,7 @@ import app.termora.Application.ohMyJson
 import app.termora.ApplicationRunnerExtension
 import app.termora.Disposable
 import app.termora.ResponseException
+import app.termora.db.Data
 import app.termora.swingCoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.channels.Channel
@@ -18,6 +19,7 @@ import kotlinx.serialization.json.long
 import okhttp3.Request
 import org.apache.commons.lang3.StringUtils
 import org.slf4j.LoggerFactory
+import kotlin.concurrent.withLock
 import kotlin.random.Random
 import kotlin.time.Duration.Companion.milliseconds
 import kotlin.time.Duration.Companion.minutes
@@ -37,7 +39,6 @@ class PullService private constructor() : SyncService(), Disposable, Application
      * 多次通知只会生效一次 也就是最后一次
      */
     private val channel = Channel<String>(Channel.UNLIMITED)
-    private val accountManager get() = AccountManager.getInstance()
     private val accountProperties get() = AccountProperties.getInstance()
 
     private suspend fun schedule() {
@@ -52,7 +53,7 @@ class PullService private constructor() : SyncService(), Disposable, Application
     }
 
     private fun synchronize(id: String) {
-        pull(id)
+        syncLock.withLock { pull(id) }
     }
 
 
@@ -70,7 +71,7 @@ class PullService private constructor() : SyncService(), Disposable, Application
         val since = 0L
         var after = StringUtils.EMPTY
         var nextSince = since
-        val limit = 2
+        val limit = 100
 
         while (true) {
             val request = Request.Builder()
@@ -86,9 +87,16 @@ class PullService private constructor() : SyncService(), Disposable, Application
                 // 如果本地不存在，并且云端已经删除，那么不需要处理
                 if (data == null && e.deleted) {
                     continue
+                } else if (data != null && e.deleted && data.deleted) { // 如果云端与本地都已经删除，那么不需要处理
+                    continue
                 }
                 if (data == null || data.version != e.version) {
+                    if (log.isInfoEnabled) {
+                        log.info("数据: {}, 本地版本: {}, 云端版本: {} 触发同步", e.objectId, data?.version, e.version)
+                    }
                     trigger(e.objectId)
+                } else if (log.isDebugEnabled) {
+                    log.debug("数据: {} 本地版本与云端版本一致", e.objectId)
                 }
             }
 
@@ -109,19 +117,26 @@ class PullService private constructor() : SyncService(), Disposable, Application
             .build()
         val response = AccountHttp.client.newCall(request).execute()
 
+        if (log.isInfoEnabled) {
+            log.info("拉取数据: {} 成功, 响应码: {}", id, response.code)
+        }
+
         // 云端数据不存在直接返回
         if (response.code == 404) {
             if (log.isWarnEnabled) {
-                log.warn("The data {} does not exist in the cloud", id)
+                log.warn("数据: {} 云端不存在，本地数据将会删除", id)
             }
+            // 云端数据不存在，那么本地也要删除
+            updateData(id, synced = true, deleted = true)
             return
         }
 
         // 没有权限拉取
         if (response.code == 403) {
             if (log.isWarnEnabled) {
-                log.warn("Data.id {} No permission to push", id)
+                log.warn("数据: {} 没有权限拉取，本地数据将会删除", id)
             }
+            updateData(id, synced = true, deleted = true)
             return
         }
 
@@ -152,31 +167,79 @@ class PullService private constructor() : SyncService(), Disposable, Application
         // 如果本地不存在，
         if (row == null) {
             // 云端已经删除，那么忽略
-            if (deleted) return
-            // 保存到本地数据库
-//            databaseManager.save(ownerId, ownerType, id, type, data, version = version, synced = true)
-            return
-        }
+            if (deleted) {
+                if (log.isInfoEnabled) {
+                    log.info("数据: {}, 类型: {} 云端已经删除，本地也不存在", id, type)
+                }
+                return
+            }
 
-        // 如果本地存在，云端已经删除，那么本地删除
-        if (deleted) {
+            if (log.isInfoEnabled) {
+                log.info("数据: {}, 类型: {} 从云端拉取成功，即将保存到本地", id, type)
+            }
+
+            // 保存到本地
+            databaseManager.save(
+                Data(
+                    id = id,
+                    ownerId = ownerId,
+                    ownerType = ownerType,
+                    type = type,
+                    data = decryptData(id, data),
+                    version = version,
+                    // 因为已经是拉取最新版本了，所以这里无需再同步了
+                    synced = true,
+                    deleted = false
+                )
+            )
+
+        } else if (deleted && row.deleted.not()) { // 如果本地存在，云端已经删除，那么本地删除
+            if (log.isInfoEnabled) {
+                log.info("数据: {}, 类型: {} 云端已经删除，本地即将删除", id, type)
+            }
             databaseManager.delete(id)
-        }
-
-        // 如果本地版本大于云端版本，那么忽略，因为需要同步
-        if (row.version > version) {
+        } else if (row.version > version) { // 如果本地版本大于云端版本，那么忽略，因为需要推送到云端
+            if (log.isInfoEnabled) {
+                log.info(
+                    "数据: {}, 类型: {}, 本地版本: {}, 云端版本: {}, 本地版本高于云端版本，即将将本地数据推送至云端",
+                    id,
+                    type,
+                    row.version,
+                    version
+                )
+            }
             // 如果没有同步标识，那么修改为代同步
             if (row.synced.not()) {
                 updateData(id, false, row.version)
             }
-            return
+        } else if (row.version < version) { // 本地版本小于云端版本，那么立即修改
+
+            if (log.isInfoEnabled) {
+                log.info(
+                    "数据: {}, 类型: {}, 本地版本: {}, 云端版本: {}, 本地版本小于云端版本，即将保存到本地",
+                    id,
+                    type,
+                    row.version,
+                    version
+                )
+            }
+
+            // 解密数据
+            databaseManager.save(
+                Data(
+                    id = id,
+                    ownerId = ownerId,
+                    ownerType = ownerType,
+                    type = type,
+                    data = decryptData(id, data),
+                    version = version,
+                    // 因为已经是拉取最新版本了，所以这里无需再同步了
+                    synced = true,
+                    deleted = false
+                )
+            )
         }
 
-        // 本地版本小于云端版本，那么立即修改
-        if (row.version < version) {
-            // 保存
-//            databaseManager.save(ownerId, ownerType, id, type, data, version = version, synced = true)
-        }
 
     }
